@@ -19,124 +19,97 @@ final class UziReaderTest extends TestCase
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateException::class);
-        $this->expectExceptionMessage("Webserver client cert check not passed");
-
         $request = new Request();
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckSSLClientFailed(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateException::class);
-        $this->expectExceptionMessage("Webserver client cert check not passed");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "failed");
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckNoClientCert(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateException::class);
-        $this->expectExceptionMessage("No client certificate presented");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckCertWithoutValidData(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateNotUziException::class);
-        $this->expectExceptionMessage("No valid UZI data found");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
         $request->server->set('SSL_CLIENT_CERT', file_get_contents(__DIR__ . '/certs/mock-001-no-valid-uzi-data.cert'));
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckCertWithInvalidSAN(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateNotUziException::class);
-        $this->expectExceptionMessage("No valid UZI data found");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
         $request->server->set('SSL_CLIENT_CERT', file_get_contents(__DIR__ . '/certs/mock-002-invalid-san.cert'));
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckCertWithInvalidOtherName(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateNotUziException::class);
-        $this->expectExceptionMessage("No valid UZI data found");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
         $request->server->set('SSL_CLIENT_CERT', file_get_contents(__DIR__ . '/certs/mock-003-invalid-othername.cert'));
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckCertWithoutIa5string(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateException::class);
-        $this->expectExceptionMessage("No ia5String");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
         $cert = file_get_contents(__DIR__ . '/certs/mock-004-othername-without-ia5string.cert');
         $request->server->set('SSL_CLIENT_CERT', $cert);
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckCertIncorrectSanData(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateException::class);
-        $this->expectExceptionMessage("Incorrect SAN found");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
         $cert = file_get_contents(__DIR__ . '/certs/mock-005-incorrect-san-data.cert');
         $request->server->set('SSL_CLIENT_CERT', $cert);
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckCertIncorrectSanData2(): void
     {
         $uzi = new UziReader();
 
-        $this->expectException(UziCertificateException::class);
-        $this->expectExceptionMessage("Incorrect SAN found");
-
         $request = new Request();
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
         $cert = file_get_contents(__DIR__ . '/certs/mock-006-incorrect-san-data.cert');
         $request->server->set('SSL_CLIENT_CERT', $cert);
 
-        $uzi->getDataFromRequest($request);
+        $this->assertNull($uzi->getDataFromRequest($request));
     }
 
     public function testCheckValidCert(): void
@@ -147,17 +120,17 @@ final class UziReaderTest extends TestCase
         $request->server->set('SSL_CLIENT_VERIFY', "SUCCESS");
         $request->server->set('SSL_CLIENT_CERT', file_get_contents(__DIR__ . '/certs/mock-011-correct.cert'));
 
-        $user = $uzi->getDataFromRequest($request);
+        $uziInfo = $uzi->getDataFromRequest($request);
 
-        $this->assertEquals('00000000', $user->getAgbCode());
-        $this->assertEquals('N', $user->getCardType());
-        $this->assertEquals('john', $user->getGivenName());
-        $this->assertEquals('2.16.528.1.1003.1.3.5.5.2', $user->getOidCa());
-        $this->assertEquals('30.015', $user->getRole());
-        $this->assertEquals('90000111', $user->getSubscriberNumber());
-        $this->assertEquals('doe-12345678', $user->getSurName());
-        $this->assertEquals('12345678', $user->getUziNumber());
-        $this->assertEquals('1', $user->getUziVersion());
+        $this->assertEquals('00000000', $uziInfo->getAgbCode());
+        $this->assertEquals('N', $uziInfo->getCardType());
+        $this->assertEquals('john', $uziInfo->getGivenName());
+        $this->assertEquals('2.16.528.1.1003.1.3.5.5.2', $uziInfo->getOidCa());
+        $this->assertEquals('30.015', $uziInfo->getRole());
+        $this->assertEquals('90000111', $uziInfo->getSubscriberNumber());
+        $this->assertEquals('doe-12345678', $uziInfo->getSurName());
+        $this->assertEquals('12345678', $uziInfo->getUziNumber());
+        $this->assertEquals('1', $uziInfo->getUziVersion());
     }
 
     public function testCheckValidAdminCert(): void
